+++ src/frontend/apps/impress/src/core/AppProvider.tsx
@@ -1,9 +1,10 @@
 import { CunninghamProvider } from '@openfun/cunningham-react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 
 import { useCunninghamTheme } from '@/cunningham';
-import { Auth } from '@/features/auth';
+import { Auth, KEY_AUTH, setAuthUrl } from '@/features/auth';
 import { useResponsiveStore } from '@/stores/';
 
 import { ConfigProvider } from './config/';
@@ -15,17 +16,19 @@ import { ConfigProvider } from './config/';
  *      - global cache duration - we decided 3 minutes
  *      - It can be overridden to each query
  */
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      staleTime: 1000 * 60 * 3,
-      retry: 1,
-    },
+const defaultOptions = {
+  queries: {
+    staleTime: 1000 * 60 * 3,
+    retry: 1,
   },
+};
+const queryClient = new QueryClient({
+  defaultOptions,
 });
 
 export function AppProvider({ children }: { children: React.ReactNode }) {
   const { theme } = useCunninghamTheme();
+  const { replace } = useRouter();
 
   const initializeResizeListener = useResponsiveStore(
     (state) => state.initializeResizeListener,
@@ -35,6 +38,27 @@ export function AppProvider({ children }: { children: React.ReactNode }) {
     const cleanupResizeListener = initializeResizeListener();
     return cleanupResizeListener;
   }, [initializeResizeListener]);
+
+  useEffect(() => {
+    queryClient.setDefaultOptions({
+      ...defaultOptions,
+      mutations: {
+        onError: (error) => {
+          if (
+            error instanceof Error &&
+            'status' in error &&
+            error.status === 401
+          ) {
+            void queryClient.resetQueries({
+              queryKey: [KEY_AUTH],
+            });
+            setAuthUrl();
+            void replace(`/401`);
+          }
+        },
+      },
+    });
+  }, [replace]);
 
   return (
     <QueryClientProvider client={queryClient}>